+++ pyscaffold/templates.py
@@ -46,14 +46,25 @@ def get_template(name):
     return string.Template(data.decode(encoding='utf8'))
 
 
-def setup(args):
+def setup_py(args):
     """
     Template of setup.py
 
     :param args: command line parameters as :obj:`argparse.Namespace`
     :return: file content as string
     """
-    template = get_template("setup")
+    template = get_template("setup_py")
+    return template.substitute(vars(args))
+
+
+def setup_cfg(args):
+    """
+    Template of setup.cfg
+
+    :param args: command line parameters as :obj:`argparse.Namespace`
+    :return: file content as string
+    """
+    template = get_template("setup_cfg")
     return template.substitute(vars(args))
 
 
@@ -287,3 +298,14 @@ def best_fit_license(txt):
     """
     ratings = {lic: levenshtein(txt, lic.lower()) for lic in licenses}
     return min(ratings.items(), key=itemgetter(1))[0]
+
+
+def namespace(args):
+    """
+    Template of __init__.py defining a namespace package
+
+    :param args: command line parameters as :obj:`argparse.Namespace`
+    :return: file content as string
+    """
+    template = get_template("namespace")
+    return template.substitute(vars(args))