+++ tests/test_structure.py
@@ -3,8 +3,8 @@
 import os
 from os.path import isdir, isfile
 
-import six
 import pytest
+import six
 from pyscaffold import runner, structure
 
 from .fixtures import nodjango_admin_mock, tmpdir  # noqa
@@ -126,3 +126,19 @@ def test_remove_from_struct():
     del_struct = {"a": 1, "c": {"a": 1}}
     res = structure.remove_from_struct(orig_struct, del_struct)
     assert res == {"b": 2, "c": {"b": 2}}
+
+
+def test_add_namespace():
+    args = ["project",
+            "-p", "package",
+            "--with-namespace", "com.blue_yonder"]
+    args = runner.parse_args(args)
+    args.namespace = runner.prepare_namespace(args.namespace)
+    struct = {"project": {"package": {"file1": "Content"}}}
+    ns_struct = structure.add_namespace(args, struct)
+    assert ["project"] == list(ns_struct.keys())
+    assert "package" not in list(ns_struct.keys())
+    assert ["com"] == list(ns_struct["project"].keys())
+    assert {"blue_yonder", "__init__.py"} == set(
+        ns_struct["project"]["com"].keys())
+    assert "package" in list(ns_struct["project"]["com"]["blue_yonder"].keys())